requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
aptdaemon==2.0.2
asgiref==3.11.0
attrs==25.1.0
autocommand==2.2.2
babel==2.17.0
bcc==0.30.0
bcrypt==4.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brlapi==0.8.6
certifi==2025.1.31
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloud-init==25.2
command-not-found==0.3
configobj==5.0.9
coverage==7.11.0
cryptography==46.0.3
cssselect==1.3.0
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
diploma-frontend @ file:///home/vladimirlesh/PycharmProjects/python_django_diploma/diploma-frontend/dist/diploma_frontend-0.6.tar.gz#sha256=5d8b4cf62782ee8d91483c92e13576d0a6639916e7f064d9e6984e054bfd65c2
distlib==0.4.0
distro==1.9.0
distro-info==1.13
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
duplicity==3.0.4
EbookLib==0.18
fastapi==0.120.2
fasteners==0.18
filelock==3.20.0
flake8==6.0.0
Flask==3.1.0
flask-cors==6.0.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
greenlet==3.2.4
gunicorn==21.2.0
gyp==0.1
h11==0.16.0
html5lib-modern==1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflect==7.3.1
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
language-selector==0.1
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.32.0
lxml==6.0.2
Mako==1.3.8.dev0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.20.1
mccabe==0.7.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.6.0
mypy==1.5.1
mypy_extensions==1.1.0
netaddr==1.3.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.47
packaging==24.2
paramiko==3.5.1
pdfminer.six==20221105
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.5.0
pluggy==1.6.0
psutil==5.9.8
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pyasyncore==1.0.2
pycairo==1.27.0
pycodestyle==2.10.0
pycparser==2.23
pycups==2.0.4
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.0.1
Pygments==2.18.0
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.10.1
PyMuPDF==1.26.6
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.1.2
pyserial==3.5
pytest==7.4.2
pytest-cov==4.1.0
python-apt==3.0.0
python-dateutil==2.9.0
python-debian==1.0.1+ubuntu1
python-dotenv==1.2.1
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.1
pyxdg==0.28
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.49.1
systemd-python==235
trash-cli==0.23.11.10
typeguard==4.4.2
types-setuptools==68.0.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.3.0
usb-creator==0.3.16
uvicorn==0.38.0
virtualenv==20.35.4
wadllib==2.0.0
webencodings==0.5.1
Werkzeug==3.1.0
wheel==0.42.0
xdg==5
xkit==0.0.0
zipp==3.21.0
